# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OKU_Prototype",
    "description": "OKU prototype: AI-personalized exam and learning assistant",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/students/me/history": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My History",
        "operationId": "my_history_api_v1_students_me_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryItemResponse"
                  },
                  "type": "array",
                  "title": "Response My History Api V1 Students Me History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/students/me/progress": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My Progress",
        "operationId": "my_progress_api_v1_students_me_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgressResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/students/me/dashboard": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My Dashboard",
        "operationId": "my_dashboard_api_v1_students_me_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/teacher/students/{student_id}/progress": {
      "get": {
        "tags": [
          "teacher"
        ],
        "summary": "Student Progress For Teacher",
        "operationId": "student_progress_for_teacher_api_v1_teacher_students__student_id__progress_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/students/{student_id}/history": {
      "get": {
        "tags": [
          "teacher"
        ],
        "summary": "Student History For Teacher",
        "operationId": "student_history_for_teacher_api_v1_teacher_students__student_id__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItemResponse"
                  },
                  "title": "Response Student History For Teacher Api V1 Teacher Students  Student Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/students/me/history": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My History",
        "operationId": "my_history_students_me_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryItemResponse"
                  },
                  "type": "array",
                  "title": "Response My History Students Me History Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/students/me/progress": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My Progress",
        "operationId": "my_progress_students_me_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgressResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/students/me/dashboard": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "My Dashboard",
        "operationId": "my_dashboard_students_me_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/teacher/students/{student_id}/progress": {
      "get": {
        "tags": [
          "teacher"
        ],
        "summary": "Student Progress For Teacher",
        "operationId": "student_progress_for_teacher_teacher_students__student_id__progress_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teacher/students/{student_id}/history": {
      "get": {
        "tags": [
          "teacher"
        ],
        "summary": "Student History For Teacher",
        "operationId": "student_history_for_teacher_teacher_students__student_id__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItemResponse"
                  },
                  "title": "Response Student History For Teacher Teacher Students  Student Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DifficultyLevel": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "DifficultyLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItemResponse": {
        "properties": {
          "test_id": {
            "type": "integer",
            "title": "Test Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "exam_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Kind"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "language": {
            "$ref": "#/components/schemas/PreferredLanguage"
          },
          "mode": {
            "$ref": "#/components/schemas/TestMode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "percent": {
            "type": "number",
            "title": "Percent"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count",
            "default": 0
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          }
        },
        "type": "object",
        "required": [
          "test_id",
          "subject_id",
          "subject_name",
          "difficulty",
          "language",
          "mode",
          "created_at",
          "percent",
          "weak_topics"
        ],
        "title": "HistoryItemResponse"
      },
...
      "PreferredLanguage": {
        "type": "string",
        "enum": [
          "RU",
          "KZ"
        ],
        "title": "PreferredLanguage"
      },
...
      "ProgressPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "percent": {
            "type": "number",
            "title": "Percent"
          }
        },
        "type": "object",
        "required": [
          "date",
          "percent"
        ],
        "title": "ProgressPoint"
      },
...
      "StudentDashboardResponse": {
        "properties": {
          "progress": {
            "$ref": "#/components/schemas/StudentProgressResponse"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryItemResponse"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "progress",
          "history"
        ],
        "title": "StudentDashboardResponse"
      },
      "StudentProgressResponse": {
        "properties": {
          "total_tests": {
            "type": "integer",
            "title": "Total Tests"
          },
          "total_warnings": {
            "type": "integer",
            "title": "Total Warnings",
            "default": 0
          },
          "avg_percent": {
            "type": "number",
            "title": "Avg Percent"
          },
          "best_percent": {
            "type": "number",
            "title": "Best Percent"
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/ProgressPoint"
            },
            "type": "array",
            "title": "Trend"
          },
          "subject_stats": {
            "items": {
              "$ref": "#/components/schemas/SubjectStat"
            },
            "type": "array",
            "title": "Subject Stats"
          }
        },
        "type": "object",
        "required": [
          "total_tests",
          "avg_percent",
          "best_percent",
          "weak_topics",
          "trend",
          "subject_stats"
        ],
        "title": "StudentProgressResponse"
      },
...
      "SubjectStat": {
        "properties": {
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "tests_count": {
            "type": "integer",
            "title": "Tests Count"
          },
          "avg_percent": {
            "type": "number",
            "title": "Avg Percent"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "subject_name",
          "tests_count",
          "avg_percent"
        ],
        "title": "SubjectStat"
      },
...
      "TestMode": {
        "type": "string",
        "enum": [
          "text",
          "audio",
          "oral"
        ],
        "title": "TestMode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}